+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### [13.27.0](https://kaos.sh/ek/13.27.0)
+
+- **`[system]`** Read sessions data from [`utmp`](https://www.man7.org/linux/man-pages/man5/utmp.5.html) file
+- **`[system]`** Added `Host` info to `SessionInfo`
+- **`[system]`** User info struct replaced by user name in `SessionInfo`
+
 ### [13.26.2](https://kaos.sh/ek/13.26.2)
 
 - **`[support]`** Improved handling empty slices with data (_deps, apps, services…_)